requirements.txt
click
google-cloud-compute